Create interface for DOSGuard
