+++ src/web/CMakeLists.txt
@@ -1,8 +1,15 @@
 add_library(clio_web)
 
 target_sources(
-  clio_web PRIVATE impl/AdminVerificationStrategy.cpp impl/ServerSslContext.cpp IntervalSweepHandler.cpp Resolver.cpp
-                   Server.cpp
+  clio_web
+  PRIVATE Resolver.cpp
+          Server.cpp
+          dosguard/DOSGuard.cpp
+          dosguard/IntervalSweepHandler.cpp
+          dosguard/WhitelistHandler.cpp
+          impl/AdminVerificationStrategy.cpp
+          impl/ServerSslContext.cpp
+          ng/Server.cpp
 )
 
 target_link_libraries(clio_web PUBLIC clio_util)
