+++ src/web/HttpSession.hpp
@@ -20,8 +20,8 @@
 #pragma once
 
 #include "util/Taggable.hpp"
-#include "web/DOSGuard.hpp"
 #include "web/PlainWsSession.hpp"
+#include "web/dosguard/DOSGuardInterface.hpp"
 #include "web/impl/HttpBase.hpp"
 #include "web/interface/ConnectionBase.hpp"
 
@@ -70,7 +70,7 @@ public:
         std::string const& ip,
         std::shared_ptr<impl::AdminVerificationStrategy> const& adminVerification,
         std::reference_wrapper<util::TagDecoratorFactory const> tagFactory,
-        std::reference_wrapper<web::DOSGuard> dosGuard,
+        std::reference_wrapper<dosguard::DOSGuardInterface> dosGuard,
         std::shared_ptr<HandlerType> const& handler,
         boost::beast::flat_buffer buffer
     )
