+++ tests/unit/CMakeLists.txt
@@ -12,7 +12,6 @@ target_sources(
           data/cassandra/ExecutionStrategyTests.cpp
           data/cassandra/RetryPolicyTests.cpp
           data/cassandra/SettingsProviderTests.cpp
-          DOSGuardTests.cpp
           # ETL
           etl/AmendmentBlockHandlerTests.cpp
           etl/CacheLoaderSettingsTests.cpp
@@ -127,11 +126,12 @@ target_sources(
           util/TxUtilTests.cpp
           # Webserver
           web/AdminVerificationTests.cpp
+          web/dosguard/DOSGuardTests.cpp
+          web/dosguard/IntervalSweepHandlerTests.cpp
+          web/dosguard/WhitelistHandlerTests.cpp
           web/impl/ServerSslContextTests.cpp
           web/RPCServerHandlerTests.cpp
           web/ServerTests.cpp
-          web/IntervalSweepHandlerTests.cpp
-          web/WhitelistHandlerTests.cpp
           # New Config
           util/newconfig/ArrayViewTests.cpp
           util/newconfig/ObjectViewTests.cpp
